requirements.txt
sseclient-py
rich
